toposerver 代码优化

+++ src/common/metadata/object_controller.go
@@ -18,7 +18,7 @@ import (
 
 // RspID response id
 type RspID struct {
-	ID int `json:"id"`
+	ID int64 `json:"id"`
 }
 
 // CreateResult create result
@@ -71,7 +71,7 @@ type UpdateGroupCondition struct {
 
 	Data struct {
 		Name  string `json:"bk_group_name"`
-		Index int    `json:"bk_group_index"`
+		Index int64  `json:"bk_group_index"`
 	} `json:"data"`
 }
